requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.5.0
mistralai>=0.0.7
loguru>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Web development
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-login>=0.6.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0

# Frontend development
nodeenv>=1.8.0

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
pylint>=3.0.0
mypy>=1.0.0

# Utilities
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.66.0
python-dateutil>=2.8.2
